backend/requirements.txt
Django==4.2.8
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.0
djongo==1.3.6
pymongo==3.12.3
Pillow==10.1.0
python-dotenv==1.0.0